poc/requirements.txt

transformers>=4.30.0          # HF Transformers 核心库 :contentReference[oaicite:0]{index=0}
accelerate>=0.20.3           # GPU/TPU 加速（可选）
sentencepiece               # 某些模型依赖 :contentReference[oaicite:1]{index=1}
huggingface-hub>=0.13.4      # 远程模型管理与下载
faker>=25.0.0
pydantic>=2.7.1
python-dotenv>=1.0.0
jinja2>=3.1.3
pandas>=2.2.0
typer>=0.9.0
rich>=13.7.0
